Make the library dependent of afb-binding
The include files for creating bindings are removed
from this project and put in an other one named
afb-binding.

This evolution has for first purpose to remove the
dependency between the projects implementing bindings
and this library. It will allow to make cool changes
in this project without enforcing recompilation of
all bindings.

Signed-off-by: José Bollo <jose.bollo@iot.bzh>