afb-sched: REWORK SCHEDULING
Add a 'mode' parameter to 'afb_sched_post_job' telling if
a thread must be started or not. It should be enough at first
and can be changed later to add more policy cases.

The previous version was starting and stopping threads,
leading to a situation in the opposite of what was first
expected: avoiding to start threads for anything.