afb-sched: Fix embraced release of evmgr
This fixes a SIGSEGV bug occurring when calling afb_[req|api]_call_sync
within the dispatch event. It occurs because the call sync release the
event manager before syncing. After that, the event manager is taken
by some other thread that expects it to be not NULL. But the dispatch
returns and put it to zero concurrently.

The solution here is to not set evmgr to NULL.