afb-extend: Don't close the shared library
When the code is removed, it leads to SIGSEGV when
remaining resources of the extensions are later cleaned up.
It has the bad effect to enforce extensions to really
free all its dependencies and that often complicates code.

The main use is to call exit at end for cleaning things
before termination so keep the shared object opened is
probably not an issue.