Fix weird pushed events when concurrency
When the data objects attached to the event are
serialized there is no special prevention of concurrency.

In the studied case the issue is probably in the function
json_object_to_json_string that in one thread starts the
serialisation and that in an other thread try to use the
not finished serialisation.

The fix use one common group for pushing events that are
then serialised.