afb-wrap-rpc: Add option "host[=[...]]" to TLS URL
When this option is added, the check of the remote hostname, has given
by it trusted certificate is no the one of the URL scheme but is
either:

 - the given name if it is not the empty string
 - not checked if the given host is the empty string

Example: The URL 'tls+tcp:192.168.0.15:34/api?host=localhost' make the
check of the certificate of the server against the name localhost with
the current system trust list.

Example: The URL 'tls+tcp:192.168.0.15:34/api?host&trust=./trusted.pem'
make the check of the certificate of the server against the given trusted
authorities but does not check the hostname of the certificate.