tls-mbed: Fix compilation issue with PSA_CRYPTO
Putting MBEDTLS_USE_PSA_CRYPTO to "y" in afb-zephyr's Kconfig.tls sets
the config value as we want in Zephyr's build config, but for some
reason, parts of MbedTLS' code do not see the value and do not compile
correctly. This leads to crashes at runtime.

Setting the value in libafb's CMakeLists solves the problem, even though
it's ugly and throws a lot of warnings during the build.